Refactor measurement_record.cc into multiple files, add docs, etc
