+++ src/simulators/detection_simulator.cc
@@ -56,10 +56,10 @@ simd_bit_table detector_samples(
 void detector_sample_out_helper(
     const Circuit &circuit, FrameSimulator &sim, size_t num_samples, bool append_observables, FILE *out,
     SampleFormat format) {
-    BatchResultWriter writer(out, num_samples, format);
+    MeasureRecordBatchWriter writer(out, num_samples, format);
     std::vector<simd_bits> observables;
     sim.reset_all();
-    writer.set_result_type('D');
+    writer.begin_result_type('D');
     simd_bit_table detector_buffer(1024, num_samples);
     size_t buffered_detectors = 0;
     circuit.for_each_operation([&](const Operation &op) {
@@ -71,7 +71,7 @@ void detector_sample_out_helper(
             }
             buffered_detectors++;
             if (buffered_detectors == 1024) {
-                writer.write_table_batch(detector_buffer, 1024 >> 6);
+                writer.batch_write_bytes(detector_buffer, 1024 >> 6);
                 buffered_detectors = 0;
             }
         } else if (op.gate->id == gate_name_to_id("OBSERVABLE_INCLUDE")) {
@@ -93,11 +93,11 @@ void detector_sample_out_helper(
         }
     });
     for (size_t k = 0; k < buffered_detectors; k++) {
-        writer.write_bit_batch(detector_buffer[k]);
+        writer.batch_write_bit(detector_buffer[k]);
     }
-    writer.set_result_type('L');
+    writer.begin_result_type('L');
     for (const auto &result : observables) {
-        writer.write_bit_batch(result);
+        writer.batch_write_bit(result);
     }
     writer.write_end();
 }
