+++ src/simulators/tableau_simulator.h
@@ -27,14 +27,14 @@
 #include "../circuit/circuit.h"
 #include "../stabilizers/tableau.h"
 #include "../stabilizers/tableau_transposed_raii.h"
-#include "measurement_record.h"
+#include "measure_record.h"
 #include "vector_simulator.h"
 
 struct TableauSimulator {
     Tableau inv_state;
     std::mt19937_64 &rng;
     int8_t sign_bias;
-    SingleMeasurementRecord measurement_record;
+    MeasureRecord measurement_record;
     bool last_correlated_error_occurred;
 
     /// Args:
@@ -43,8 +43,7 @@ struct TableauSimulator {
     ///     sign_bias: 0 means collapse randomly, -1 means collapse towards True, +1 means collapse towards False.
     ///     record: Measurement record configuration.
     explicit TableauSimulator(
-        size_t num_qubits, std::mt19937_64 &rng, int8_t sign_bias = 0,
-        SingleMeasurementRecord record = SingleMeasurementRecord());
+        size_t num_qubits, std::mt19937_64 &rng, int8_t sign_bias = 0, MeasureRecord record = MeasureRecord());
 
     /// Samples the given circuit in a deterministic fashion.
     ///
